requirements.txt
diffusers
matplotlib
numpy
opencv_python
Pillow
Pandas
requests
setuptools
scikit-image
transformers
torch
torchvision
torchmetrics
